Adding optional gzip output capability (new -g or --gzip option for msconvert)
Tech note on doing this the boost-y way:  It's *almost* a very simple matter to add gzip output.  I say "almost" because index writing is complicated by the fact that tellp() isn't implemented for the gzip output filter used with boost::iostreams::filtering_ostream.

Fair enough, the idea of tellp() is ambiguous anyway (was that bytes into or out of the compressor?).  There's a "counter" filter available for counting the bytes passed into the stream, so we just add that to the filter stack, and in XMLWriter::Impl::position() use dynamic_cast<> to determine whether we're actually writing to an ostream (use tellp()) or a filtering_ostream (use the counting filter).

Oops, the provided "counter" filter counts using int instead of stream_position.  No worries, we just copy the code and make a new version that counts with stream_position (and drops the line counting we didn't want anyway).  Voila!

In the end, boost makes it pretty tidy.