Add "MayOverlap" to peaks in AllChromatogramsGraph, so only SRM and cached peaks need to be checked for overlap and summation.
Save "auto-show" of AllChromatogramsGraph window as a setting.  Add a checkbox to ImportResultsDlg to show the graph, and a button in the status bar of the main window, next to the progress bar.

Remove transparent gradient from peaks in AllChromatogramsGraph for better rendering performance in ZedGraph.

Remove peaks that fall below the threshold after scaling the intensity axis.  This improves rendering performance and makes the graph look cleaner.

Attempt to do finer-grained cancellation of import.  It could still be better.

Move support for fine-grained control of DotTrace profiler to Util from TestUtil, so it's available from anywhere in Skyline.