Merging my branch for vendor reader performance measurement in Skyline chromatogram extraction.  A couple of things here probably demand a clean before you build again.
Replace DotNetZip with latest version to work around issues with large zipfiles - was using 32 bit seek, which meant any zipfile with entries beginning past the 32 bit boundary would fail.  This was a DotNetZip API-breaking revision so a couple of our source files changed with it.

Enabled mz5 as a recognized data input format for Skyline.

Allow user to type a full pathname into the textbox in OpenDataSourceDialog.  Formerly gave a rather confusing error message when you did this - "you must select a file", which of course you thought you just had.

Fixed a bug when dealing with missing .skyd files and the need to re-import Results (instrument output) files, which would claim .d "files" couldn't be located (because they're directories, not files - now we check for both).

When a results file actually is missing, give user a more specific message - "the following files are missing" as opposed to "some files are missing".

Added the "TestPerf" project for vendor-vs-mz5 file read performance measurement, but without all tests commented out since they're not interesting in day-to-day use and can really slow things down.  This also introduces the idea of persistent test files - things you'd like to download just once (which we already supported) and unzip just once in the downloads area for repeated use (which we did not).  Normally that would be some sort of read-only data like mass spec data files.  Note that Bruker files are commented out here - for Bruker it's still DLL hell under MSTest and I finally punted and did this tests manually.