Fix two Skyline exceptions. 1) ArgumentNullException in GetRearrangedMods (RefinementSettings.cs) https://skyline.gs.washington.edu/labkey/announcements/home/issues/exceptions/thread.view?rowId=9076 We would hit this exception if you had a peptide with explicit modifications, but no static modifications.
2) ArgumentException in Path.GetFileNameWithoutExtension in SkylineGraphs.cs
If the file had a sample name appended with a "|", we would throw an exception when I tried to call Path.GetFileNameWithoutExtension.  I need to call "SampleHelp.GetFilePath" to strip off everything after the "|"
Next version, I will change ChromFileInfo.FilePath into an object so that I don't make this type of mistake again (this will also help with Chorus integration, because the file path will sometimes be a URL).