Skyline: Fix problem where, when importing results, new rows do not appear automatically in the Results Grid.
This change also changes the way that Live Reports send change notifications to the DataGridView.
They used to notice when ColumnDescriptor.GetValue() was called, and use the class RowItemList to add a listener to the Document, and then send PropertyChanged notifications.
Instead, those ListChangedEvents get fired by BindingListView.RowSourceListChanged.

Added DocumentGridTest which ensures that the ResultsGrid and the DocumentGrid both show the correct number of rows while replicates are being imported.