Add TestInteractiveTool, a new Skyline tool that is run by ToolServiceTest to test the ToolService API for interactive tools.
Do a better job of shutting down the communication channel so it doesn't take 10-15 seconds to shut down a service.

Change ToolInstaller so you can install a tool from a directory in addition to a zip file.  Using the test framework, a zip file is already expanded before we install the tool.

Change to using properties for ISkylineTool interface.  I didn't know how to do that in WCF previously.