Add support for interactive Skyline tools that can ask Skyline for updated reports and select items in the Skyline window.
Details for some of the incremental changes on the branch:

Change ToolInstaller so you can install a tool from a directory in addition to a zip file.  Using the test framework, a zip file is already expanded before we install the tool.

Fix one character offset for link in Skyline's about box.

Add TestInteractiveTool, a new Skyline tool that is run by ToolServiceTest to test the ToolService API for interactive tools.

Also incorporate some new sorting criteria for transitions to match new VS 2013 branch.

Fixed a Zedgraph bug with y-axis zooming. It wasn't respecting the no-vpan setting.