Skyline: adding support for generalized small molecules.  These are extensive changes.
Users will notice:
A new type of node in the targets tree, with its own icon, for Molecules.  Molecules can be specified by chemical formula and charge, or by mass and charge.

Developers will notice:

Formerly we assumed all molecules were peptides, and treated them as synonyms.  Now we have the idea of "Molecule", which is the base class for "Peptide" and "SmallMolecule".

Some of the most common document properties have been renamed to support this new generality.  Some examples:

Example 1:
        public enum Level { PeptideGroups, Peptides, TransitionGroups, Transitions }
becomes
        public enum Level { MoleculeGroups, Molecules, TransitionGroups, Transitions }
Example 2: SrmDocument.TransitionCount formerly dealt with peptides only (since that's all we had), for that you now want SrmDocument.PeptideTransitionCount.  SrmDocument.TransitionCount now returns the count of transitions of any type - peptide or small molecule.
Example 3: SrmDocument.Peptides formerly returned a list of all PeptideDocNodes.  Now it only returns those that are actually peptides.  SrmDocument.Molecules returns the comprehensive list of peptides and small molecules together.

SkylineTester has a new checkbox for enabling a special test mode wherein documents in most existing tests automatically have a special small molecule node added to them, to increase our coverage of this new functionality.  When this is enabled and multiple loops are indicated, the special node behaviour happens on every other loop.

Caveats for users: 
There are still a number of awkward fits for small molecules in a formerly proteomics-centric product, particularly in the area of reporting (for small molecules, for example, for "Sequence" we may substitute the molecular formula).

Caveats for developers: 
The renaming process is incomplete - we didn't want to churn absolutely everything in one commit.  For example, what you would expect to be called MoleculeDocNode is still called PeptideDocNode - this is confusing since SrmDocument.Molecules returns a list of PeptideDocNodes.  Expect this to change to a more rational scheme eventually.