Skyline: make "not a valid chemical formula" error message in Insert Transition List more helpful by including a list of supported symbols.
We'd already fixed the problem where the error was an exception rather than a tidy message in the paste window, but where the error message was formerly
   "The expression 'C12D3H9' is not a valid chemical formula."
it's now
   "The expression 'C12D3H9' is not a valid chemical formula. Supported chemical symbols include: H H' O O" O' N N' C C' S P Se Li F Na Cl K Ca Fe Ni Cu Zn Br Mo Ag I Au Hg B As Cd Cr Co Mn Mg Si"
which fits in the existing space just fine.

This is in response to https://skyline.gs.washington.edu/labkey/announcements/home/issues/exceptions/thread.view?entityId=cf818116-87a3-1032-a707-da20258204f4&_anchor=19034#row:19034.  =