AbstractViewContext.cs: When exporting from the document grid, any quickfilter, or sort was not being applied.
GroupComparer.cs:
The "Replicate Count" column is supposed to show the number of replicates that were included in the calculation (that is, were not thrown out for missing data for any transition). If there were technical replicates, this column was showing the number of distinct subjects.

ToolStoreDialog.cs:
Fix exception