Skyline (small molecules isotope labels branch):  move much small molecule information away from the ID nodes to the DocNodes, since users manipulate small molecules in a way that they do not with peptides (with peptides we can generate nodes in ways that aren't possible with small molecules).
This is just a checkpoint commit - at this stage the molecule ion information is completely detached from the precursors.  Our intent is that the molecule show the ion info from the first precursor in its transition group list (and that molecules with no precursors are not allowed to exist).  But all tests do pass.