Create a new class AlertDlg, which both MessageDlg and MultiButtonMsgDlg. The big motivator for this was that we needed to add a "More Info" button to "MultiButtonMsgDlg".
AlertDlg uses a FlowLayoutPanel to display the buttons.  You can call the method "AddButton(DialogResult, string)" to add a button to the button bar.

AlertDlg also knows how to translate "MessageBoxButtons" enums into sets up buttons to go on the button bar.  In the future, all uses of "MessageBox" will be replaced with some form of AlertDlg.